API BREAK: fix: typedef 'ssize_t' to 'SSIZE_T' on Windows

Closes #1804

See merge request graphviz/graphviz!2433

+++ lib/gvpr/gvpr.h
@@ -28,7 +28,8 @@ extern "C" {
 
 #include "cgraph.h"
 #ifdef _MSC_VER
-typedef int ssize_t;
+#include <BaseTsd.h>
+typedef SSIZE_T ssize_t;
 #endif
 
 /* Bits for flags variable in gvprstate_t.
